better onHeapTableIndex
